% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_DR2Nagelkerke.R
\name{Calculate_DR2Nagelkerke}
\alias{Calculate_DR2Nagelkerke}
\title{Calculating the increase of pseudo-R2 of Nagelkerke due to PRS}
\usage{
Calculate_DR2Nagelkerke(SNPs.list, sumGWAS, amb.remove = FALSE, Cov = NULL,
  path.to.plink.files, bfile, P.threshold = 1, cov.names = NULL,
  better.coef = NULL, output.file = "output")
}
\arguments{
\item{SNPs.list}{A dataframe with one column,indicating the SNPs belonging to the geneset, in the format chr:position.
This is the return of function \code{Map_genestoSNPs}.}

\item{sumGWAS}{A dataframe with a minimum of four columns with names "snpid", "a1", "p","effect", taken from the summary
statistics of the GWAS discovery sample. SNPs in the "snpid" column has to be identified as chr:pos (such as 4:103334).
The "effect" column indicates the effect attributed to allele "a1", as logOR (beta coefficient).}

\item{amb.remove}{A logical indicating if SNPs with strand ambiguity (i.e. A<->T and C<->G) have to be removed from sumGWAS
before the estimation of PRS. If \code{TRUE}, the sumGWAS has to include an additional column with name "a2" to indicate
the non-effect allele at each SNP.}

\item{Cov}{A dataframe with a column "IID"  for sample identification and covariates with specific names.}

\item{path.to.plink.files}{String with the full path to the three PLINK binary files (.bed, .bim, .fam)}

\item{bfile}{A string indicating the name of the PLINK binary file with data of the GWAS target sample, without extension.}

\item{P.threshold}{A number indicating the P thershold of the GWAS discovery sample to include an SNP in the PRS model.}

\item{cov.names}{An array with the names of covariates from the \code{Cov} dataframe to be included in the model.}

\item{better.coef}{An array of two numbers, indicating disease prevalence and proportion of cases in the target sample.
Only needed to calculate a better coefficient of determination than Nagelkerke's pseudo-R2, interpetable on the
liability scale and adjusted for ascertainment bias. See Lee et al (2012).}

\item{output.file}{A string indicating the name of the output file to write the main data of the analysis. This file
is generated in the path defined in \code{path.to.plink.files}}
}
\value{
A list with four elements. A number representing the increase in pseudo-R2 of Nagelkerke, an array of SNPs
  included in the PRS model, the summary of the logistic regression model (generated by the \code{lrs} function at
  the \code{rms} package), and a better coefficient of determination, if requested.
}
\description{
\code{Calculate_DR2Nagelkerke} calculates the increase in pseudo-R2 of Naglekerke between a full logistic regression model,
   considering covariates and the PRS, and a model considering only covariates. It also creates
   an array with the list of SNPs in the PRS model, identified by chromosome and position (ex. 4:103188709).
   It also generates a file with the main data of the analysis, including the proccess of SNP clumping, as well as
   the value of increase in pseudo-R2 of Nagelkerke, and the OR (and 95 percent C.I.) of the standardized score. Optionally,
   it also calculates a better coefficient of determination than Nagelkerke's pseudo-R2, interpetable on the
   liability scale and adjusted for ascertainment bias. See Lee et al (2012).
}
