% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_permutationSets.R
\name{Generate_permutationSets}
\alias{Generate_permutationSets}
\title{Generating permutation sets of random SNPs with similar properties than the actual SNP set}
\usage{
Generate_permutationSets(input.perm, sumGWAS, path.to.plink.files,
  bfile = "tmp", n.perm)
}
\arguments{
\item{input.perm}{An array of SNPs included in the PRS model, usually, the second element of the list generated
by \code{Calculate_DR2Nagelkerke}.}

\item{sumGWAS}{A dataframe with a minimum of four columns with names "snpid", "a1", "p","effect", taken from the summary
statistics of the GWAS discovery sample. SNPs in the "snpid" column has to be identified as chr:pos (such as 4:103334).
The "effect" column indicates the effect attributed to allele "a1", as logOR (beta coefficient).}

\item{path.to.plink.files}{String with the full path to the three PLINK binary files (.bed, .bim, .fam).}

\item{bfile}{A string indicating the name of the PLINK binary file with data of the GWAS target sample, without extension.
The tmp files generated by the Calculate_DR2Nagelkerke are used by default.}

\item{n.perm}{An integer indicating how many random SNP sets have to be generated.}
}
\value{
A dataframe with the number of random SNP sets especified by \code{n.perm} matching the original
 SNP set in several properties.
}
\description{
\code{Generate_permutationSets} generates random sets of SNPs with the same size of the actual SNP set,
  aproximately matching the actual SNP set in several properties, such as gene density and distance to nearest
  gene, minor allele frequency, and linkage disequilibrium, based on the 1000 genomes European population, taking
  data from SNPsnap (Pers et al., 2015).
}
\details{
The extended Major Histocompatibility Complex (chr6, from 25912984 to 33290793 in hg19) was excluded due to its
  complex linkage disequilibrium structure.

The SNPs are matched based on deciles except for gene counts, due to the excess of SNPs with low gene counts. Thus,
  80.3 percent of SNPs present 0-4 gene counts. The classes to match were 0, 1, 2, 3, 4, and two additional classes, evenly distributed.
  Those SNPs not present in the SNPsnap data set or without any matched SNP in the SNPsnap data set are included in
  each permutation, to get the same number of SNPs than in the original SNP set.
}
