% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Plot.R
\name{Generate_Plot}
\alias{Generate_Plot}
\title{Generating the histogram with the distribution of permuted pseudo-R2 values}
\usage{
Generate_Plot(permDR2 = permDR2, DR2 = DR2)
}
\arguments{
\item{permDR2}{Array with pseudo-R2 values from permutations. It is the second element of the list generated as output by
the \code{Calculate_permutationSignificance} function of this package.}

\item{DR2}{A number corresponding to the value of the increase in pseudo-R2 of Naglekerke of the actual SNPs set. It is the
first element of the list generated by the \code{calculate_DR2Nagelkerke} function of this package.}
}
\value{
A plot showing the distribution of pseudo-R2 for permutations and the actual SNPs set.
}
\description{
\code{Generate_Plot} returns a histogram representing the distribution of the Nagelkerke's pseudo-R2 values for the permuted SNPs sets.
  A vertical red line with the value for the actual SNPs set is also shown. Its purpose is to get a fast visualization of the results.
}
