% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_genestoSNPs.R
\name{Map_genestoSNPs}
\alias{Map_genestoSNPs}
\title{Selecting SNPs located in a specific list of genes}
\usage{
Map_genestoSNPs(geneset, bfile, path.to.plink.files, extra.kb = c(0, 0),
  output = "missing_genes.txt")
}
\arguments{
\item{geneset}{List of genes from gene set, HGNC symbol or HGNC_ID, no header.}

\item{bfile}{A string indicating the name of the PLINK binary file with data of the GWAS target sample, without file extension.}

\item{path.to.plink.files}{String with the full path to the three PLINK binary files (.bed, .bim, .fam)}

\item{extra.kb}{An array of two numbers indicating the extra bases in kb to add at the 5' and 3' edges of the
coding region for mapping SNPs to genes.}

\item{output}{A string indicating the name of the output file to write the list of genes not identified. This file
is generated in the path defined in \code{path.to.plink.files}.}
}
\value{
A dataframe of one column with the SNPs mapping your genes, identified by chromosome and position (ex. 4:103188709).
}
\description{
\code{Map_genestoSNPs} takes a specified gene set and returns a list of SNPs from a genotype data set located within these genes
 including n bases at both edges of the coding region. It also generates a file with a list of genes from the gene set
 not found in the hg19 (NCBI37.3) version of the human genome. This is useful to check for the correct name.
}
